data/Fissures/Fissure route/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
catboost
joblib
rich
python-pptx
pymannkendall
astral
statsmodels
